refactor: extract helper functions into separate files
